+++ src/test/java/org/braekpo1nt/mctmanager/ui/sidebar/MockFastBoardWrapper.java
@@ -1,5 +1,6 @@
-package org.braekpo1nt.mctmanager.ui;
+package org.braekpo1nt.mctmanager.ui.sidebar;
 
+import org.braekpo1nt.mctmanager.ui.sidebar.FastBoardWrapper;
 import org.bukkit.entity.Player;
 
 public class MockFastBoardWrapper extends FastBoardWrapper {
